fix(*): windows test timestamp error.

+++ src/mynteye/device/channel/channels.cc
@@ -533,7 +533,7 @@ void Channels::CheckTimeStampLimmit(mynteye::ImuPacket2 &packet) {
     }
     if (is_nearly_before_timestamp_limmit_ > 0) {
       is_nearly_before_timestamp_limmit_--;
-      if (fabs(current_datum_ - segment2.timestamp) > (uint64_t)(42949672960/2)) {  // NOLINT
+      if ((uint64_t)(fabs(current_datum_ - segment2.timestamp)) > (uint64_t)(42949672960/2)) {  // NOLINT
         segment2.timestamp -= 42949672960;
       }
     }
